// src/interfaces/userinfo-data.ts
export interface UserInfoData {
  id: number;
  user_id: number;
  name: string;
  document: string;
  document_type: string;
  email: string
  phone: string;
  nationality: string;
  cep: string;
  city: string;
  state: string;
  neighborhood: string;
  street: string;
  location: string;
  flags: null;
  created_at: string;
  updated_at: string;
  house_number: string;
  mothersName: string;
  birthday: string;
  active: string;
  uuid: string;
}